implement order creation flow

+++ WebApi/Core/DTOs/OrderDTOs/OrderItemDto.cs
@@ -4,7 +4,7 @@
     {
         public long Id { get; set; }
         public long OrderId { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string? ProductName { get; set; }
         public int Quantity { get; set; }
         public decimal Price { get; set; }
@@ -13,9 +13,9 @@
     public class OrderItemCreateDto
     {
         public long OrderId { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string? ProductName { get; set; }
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
         public decimal Price { get; set; }
     }
